+++ src/app/services/user.service.ts
@@ -29,7 +29,7 @@ export class UserService {
       return this.http.get<any>(this.accountUrl).pipe(
           tap(jsonAccount =>
               this.currentAccount = jsonAccount.user ?
-                  {user: new User(jsonAccount.user), numberOfDevices: jsonAccount.numberDevices } : undefined ),
+                  {user: new User(jsonAccount.user), numberOfDevices: jsonAccount.numberOfDevices } : undefined ),
           map(_ => this.currentAccount ? this.currentAccount.user : undefined));
     }
   }
@@ -43,7 +43,7 @@ export class UserService {
             this.currentAccount = jsonAccount.user ?
                 new AccountInfo({
                   user: new User(jsonAccount.user),
-                  numberOfDevices: jsonAccount.numberDevices
+                  numberOfDevices: jsonAccount.numberOfDevices
                 }) : undefined;
             this.behaviorSubItem.next(this.currentAccount);
           }),